Adds a newline at the end of file when required
Adds a newline at the end of file when required. A lot of us gets error (by CI) while pushing codes. It's annoying to add newlines manually (if we miss somewhere). This script will ease it out.
Importance of newline at the end of file : https://unix.stackexchange.com/questions/18743/whats-the-point-in-adding-a-new-line-to-the-end-of-a-file#answers